refactor(docs): shared form fields and check UI constants
- Add TextField/TextareaField/FieldError in components/ui/form-field
- Use formatAge and STATUS_LABELS from lib in check result views
- Wire contact dialog and OG components to shared helpers

Made-with: Cursor
