+++ docs/src/components/check/check-results.tsx
@@ -8,29 +8,11 @@ import { Badge } from "@/components/ui/badge";
 import { LinkText } from "@/components/ui/link-text";
 import { Separator } from "@/components/ui/separator";
 import { STATUS_CONFIG } from "@/lib/check";
+import { formatAge } from "@/lib/date";
 import type { CheckTranslations } from "@/locales/en/check";
 import type { TranslatedCheckResponse } from "@/types/check";
 
 import { ResultSection } from "./result-section";
-
-const formatAge = (iso: string, t: CheckTranslations["time"]): string => {
-  const days = Math.floor(
-    (Date.now() - new Date(iso).getTime()) / (1000 * 60 * 60 * 24),
-  );
-  if (days === 0) {
-    return t.today;
-  }
-  if (days === 1) {
-    return t.yesterday;
-  }
-  if (days < 30) {
-    return t.daysAgo.replace("{days}", String(days));
-  }
-  if (days < 365) {
-    return t.monthsAgo.replace("{months}", String(Math.floor(days / 30)));
-  }
-  return t.yearsAgo.replace("{years}", String(Math.floor(days / 365)));
-};
 
 export const CheckResults = ({
   data,
